feat(facades,unix,fb): sellar arch/task y socketpair + MAP_SHARED fb0
Cierra prep X11 parcial: facades ir0 sin includes arch/sched, socketpair
AF_UNIX y gate MAP_SHARED /dev/fb0 con smokes KTM.

Cambios funcionales:
- arch_config/cpu/io bajo includes/ir0; arch_port/task/context sellados.
- arch-guard: facade-no-arch/sched-include.
- sys_socketpair + sock_stream_socketpair; smoke-socketpair.
- Tag FB_MAP_SHARED_OK; smoke-fb-map-shared.
- Docs ROADMAP/BACKLOG/KTM/DECOUPLING.

Tests:
- arch-guard OK; tests/host 21/21; smoke-socketpair y smoke-fb-map-shared OK.

Archivos modificados:
- Documentation/BACKLOG_REMAINING.md
- Documentation/DECOUPLING.md
- Documentation/KTM.md
- Documentation/ROADMAP.md
- Makefile
- arch/common/arch_config.h
- arch/common/arch_interface.h
- arch/common/arch_portable.h
- includes/ir0/arch_port.h
- includes/ir0/context.h
- includes/ir0/sock_stream.h
- includes/ir0/task.h
- kernel/sock_stream.c
- kernel/syscalls/mm_syscalls.c
- kernel/syscalls/socket_syscalls.c
- kernel/syscalls/socket_syscalls.h
- kernel/syscalls/syscall_dispatch.c
- sched/task.h
- scripts/architecture_guard.py

Archivos nuevos:
- includes/ir0/arch_config.h
- includes/ir0/arch_cpu.h
- includes/ir0/arch_io.h
- tests/ktm/userdev/ktm_fb_map_shared_case.c
- tests/ktm/userdev/ktm_socketpair_case.c
