+++ arch/common/arch_portable.h
@@ -1,423 +1,3 @@
 /* SPDX-License-Identifier: GPL-3.0-only */
-/**
- * IR0 Kernel — Core system software
- * Copyright (C) 2025  Iván Rodriguez
- *
- * This file is part of the IR0 Operating System.
- * Distributed under the terms of the GNU General Public License v3.0.
- * See the LICENSE file in the project root for full license information.
- *
- * File: arch_portable.h
- * Description: IR0 kernel source/header file
- */
-
-// ===============================================================================
-// IR0 KERNEL PORTABLE ARCHITECTURE INTERFACE
-// ===============================================================================
-// This file provides a portable interface for all supported architectures
-// All architecture-specific code should go through this interface
-
 #pragma once
-
-#include <stdint.h>
-#include <stddef.h>
-#include "arch_config.h"
-
-typedef uintptr_t arch_addr_t;
-typedef uintptr_t arch_size_t;
-typedef uint32_t arch_irq_t;
-typedef uint32_t arch_flags_t;
-
-
-/**
- * Initialize architecture-specific memory management
- */
-void arch_memory_init(void);
-
-/**
- * Get the start address of available physical memory
- */
-arch_addr_t arch_get_memory_start(void);
-
-/**
- * Get the end address of available physical memory
- */
-arch_addr_t arch_get_memory_end(void);
-
-/**
- * Allocate a physical page
- */
-arch_addr_t arch_alloc_page(void);
-
-/**
- * Free a physical page
- */
-void arch_free_page(arch_addr_t addr);
-
-/**
- * Map virtual address to physical address
- */
-int arch_map_page(arch_addr_t virt, arch_addr_t phys, arch_flags_t flags);
-
-/**
- * Unmap a virtual page
- */
-int arch_unmap_page(arch_addr_t virt);
-
-/**
- * Get page size for current architecture
- */
-size_t arch_get_page_size(void);
-
-/**
- * Initialize architecture-specific interrupt system
- */
-void arch_interrupt_init(void);
-
-/**
- * Compatibility alias for architecture IRQ initialization.
- */
-void arch_irq_init(void);
-
-/**
- * Unmask architecture-specific IRQ lines needed for boot devices.
- */
-void arch_boot_irq_unmask(void);
-
-/**
- * Enable interrupts globally
- */
-void arch_enable_interrupts(void);
-
-/**
- * Disable interrupts globally
- */
-void arch_disable_interrupts(void);
-
-/**
- * Register an interrupt handler
- */
-int arch_register_irq(arch_irq_t irq, void (*handler)(void));
-
-/**
- * Unregister an interrupt handler
- */
-int arch_unregister_irq(arch_irq_t irq);
-
-/**
- * Send end-of-interrupt signal
- */
-void arch_eoi(arch_irq_t irq);
-
-/**
- * Read from I/O port (x86) or MMIO address (ARM)
- */
-uint8_t arch_io_read8(arch_addr_t addr);
-
-/**
- * Write to I/O port (x86) or MMIO address (ARM)
- */
-void arch_io_write8(arch_addr_t addr, uint8_t value);
-
-/**
- * Read 16-bit from I/O port or MMIO address
- */
-uint16_t arch_io_read16(arch_addr_t addr);
-
-/**
- * Write 16-bit to I/O port or MMIO address
- */
-void arch_io_write16(arch_addr_t addr, uint16_t value);
-
-/**
- * Read 32-bit from I/O port or MMIO address
- */
-uint32_t arch_io_read32(arch_addr_t addr);
-
-/**
- * Write 32-bit to I/O port or MMIO address
- */
-void arch_io_write32(arch_addr_t addr, uint32_t value);
-
-
-/**
- * Halt the CPU (low power mode)
- */
-void arch_cpu_halt(void);
-
-/**
- * Enter architecture-specific idle wait instruction.
- */
-void arch_cpu_idle(void);
-
-/**
- * Platform final actions after coordinated shutdown (do not return).
- * Distinct from arch_cpu_idle(): these stop the machine / CPU permanently.
- */
-void arch_system_halt(void) __attribute__((noreturn));
-void arch_system_reboot(void) __attribute__((noreturn));
-void arch_system_poweroff(void) __attribute__((noreturn));
-
-/**
- * Get CPU ID (APIC ID)
- */
-uint32_t arch_get_cpu_id(void);
-
-/**
- * Get number of CPUs
- */
-uint32_t arch_get_cpu_count(void);
-
-/**
- * Get CPU vendor string
- * @vendor_buf: Buffer to store vendor string (must be at least 13 bytes)
- * Returns: 0 on success, -1 on failure
- */
-int arch_get_cpu_vendor(char *vendor_buf);
-
-/**
- * Get CPU signature (family, model, stepping)
- * @family: Output for CPU family
- * @model: Output for CPU model
- * @stepping: Output for CPU stepping
- * Returns: 0 on success, -1 on failure
- */
-int arch_get_cpu_signature(uint32_t *family, uint32_t *model, uint32_t *stepping);
-
-/**
- * Get CPUID maximum leaf (EAX from CPUID.0)
- */
-int arch_get_cpuid_max_leaf(uint32_t *max_leaf);
-
-/**
- * Get CPU brand string from silicon (CPUID 0x80000002-0x80000004)
- * @buf: Buffer, at least 49 bytes
- * @size: Buffer size
- */
-int arch_get_cpu_brand_string(char *buf, size_t size);
-
-/**
- * Get CPU feature bits from CPUID.1 (EDX, ECX)
- */
-int arch_get_cpu_feature_bits(uint32_t *out_edx, uint32_t *out_ecx);
-
-/**
- * Get CLFLUSH line size in bytes (from CPUID.1)
- */
-uint32_t arch_get_cpu_clflush_size(void);
-
-/**
- * Switch to user mode (if supported)
- */
-void arch_switch_to_user(arch_addr_t entry, arch_addr_t stack);
-
-/*
- * arch_switch_to_user_task - Enter ring 3 with full task register state.
- * Linux ret_from_fork analogue for syscall-block resume paths.
- */
-struct task;
-void arch_switch_to_user_task(const struct task *task);
-
-/*
- * arch_first_context_switch - First transfer from idle/boot into @next.
- * Does not return on success (iretq / EL drop / cooperative switch).
- * ISA details live in arch backends; portable sched must not embed iretq.
- */
-struct process;
-void arch_first_context_switch(struct process *next);
-
-/*
- * arch_set_fs_base - Set x86-64 FS base (TLS) for the running hardware thread.
- * Used by sys_arch_prctl(ARCH_SET_FS). No-op on non-x86 builds.
- */
-void arch_set_fs_base(uint64_t base);
-
-/*
- * arch_set_tls - Portable TLS base install (x86: FS base; other arch: stub).
- * Prefer this from portable code instead of CPUID/MSR details.
- */
-static inline void arch_set_tls(uint64_t base)
-{
-	arch_set_fs_base(base);
-}
-
-/*
- * arch_tls_invalidate - Drop cached TLS view after task switch prep (no-op MSR).
- */
-static inline void arch_tls_invalidate(void)
-{
-}
-
-/*
- * W10 multi-arch note (2026-07):
- * - TLS: use arch_set_tls / arch_tls_invalidate from portable paths (done).
- * - MM activate / TLB / irq save: use arch_mm_* / arch_tlb_* / arch_irq_save
- *   (F8-facade-mm). VA indices + PTE decode: arch_mm_va_indices / arch_mm_pte_*
- *   (Pack C). Table alloc/create path still mostly in mm/paging.c.
- * - Trap/context: x86-64 IDT + switch_x64.asm stay arch-local; portable code
- *   uses arch_* facades only (no CPUID in kernel/syscalls).
- */
-
-/**
- * Activate address-space root (x86: CR3; ARM64: TTBR0). Neutral name — no ISA in callers.
- */
-void arch_mm_activate(uintptr_t root);
-
-/** Read current address-space root from hardware. */
-uintptr_t arch_mm_current_root(void);
-
-/** Invalidate one VA in the local TLB. */
-void arch_tlb_invalidate_page(uintptr_t va);
-
-/** Invalidate all non-global TLB entries (local CPU). */
-void arch_tlb_invalidate_all(void);
-
-/**
- * Save IRQ state and disable IRQs. Restore with arch_irq_restore.
- * Portable replacement for pushfq/cli copies in kernel/sched.
- */
-unsigned long arch_irq_save(void);
-
-/** Restore IRQ state from arch_irq_save. */
-void arch_irq_restore(unsigned long flags);
-
-/**
- * MM control registers behind a neutral API (W10b partial).
- * ctrl0: x86 CR0; ARM64 SCTLR_EL1 (bit0 M ≈ paging enabled).
- * ctrl1: x86 CR4; ARM64 reserved/0 until a portable need exists.
- */
-uint64_t arch_mm_read_ctrl0(void);
-void arch_mm_write_ctrl0(uint64_t value);
-uint64_t arch_mm_read_ctrl1(void);
-
-/**
- * 4-level VA indices (9 bits each) for 4 KiB granules — x86-64 and aarch64.
- * idx[0]=L0/PML4, idx[1]=L1/PDPT, idx[2]=L2/PD, idx[3]=L3/PT.
- */
-void arch_mm_va_indices(uintptr_t va, size_t idx[4]);
-
-/** PTE/table descriptor decode (ISA-local present / large / phys). */
-int arch_mm_pte_present(uint64_t e);
-int arch_mm_pte_large(uint64_t e);
-uintptr_t arch_mm_pte_phys(uint64_t e);
-
-/**
- * Build non-leaf table PTE (present+RW[+user]; never NX on table levels).
- * Leaf PTE: present + low 12 flag bits; NX when exec==0 (x86).
- */
-uint64_t arch_mm_make_table_pte(uintptr_t phys, int user);
-uint64_t arch_mm_make_leaf_pte(uintptr_t phys, uint64_t flags12, int exec);
-/** OR user-accessible bit onto an existing table descriptor. */
-void arch_mm_pte_set_user(uint64_t *e);
-
-/*
- * arch_get_fs_base - Read x86-64 FS base MSR.
- */
-uint64_t arch_get_fs_base(void);
-
-/**
- * Get current CPU mode
- */
-uint32_t arch_get_cpu_mode(void);
-
-/**
- * Initialize architecture-specific timer
- */
-void arch_timer_init(void);
-
-/**
- * Initialize architecture-specific syscall entry plumbing.
- */
-void arch_syscall_init(void);
-
-/**
- * Get current timer value
- */
-uint64_t arch_timer_read(void);
-
-/**
- * Set timer frequency
- */
-void arch_timer_set_frequency(uint32_t hz);
-
-/**
- * Get timer frequency
- */
-uint32_t arch_timer_get_frequency(void);
-
-/**
- * Get fault address (for page faults, etc.)
- */
-arch_addr_t arch_get_fault_address(void);
-
-/**
- * Get fault type
- */
-uint32_t arch_get_fault_type(void);
-
-/**
- * Get fault error code
- */
-uint32_t arch_get_fault_error(void);
-
-/**
- * Dump CPU registers
- */
-void arch_dump_registers(void);
-
-/**
- * Early architecture initialization
- * Sets up core architecture features (GDT/TSS for x86, etc.)
- */
-void arch_early_init(void);
-
-/**
- * Initialize architecture-specific interrupt system
- * Sets up interrupt tables and controllers (IDT/PIC for x86, GIC for ARM, etc.)
- */
-void arch_interrupt_init(void);
-
-/**
- * Late architecture initialization
- */
-void arch_late_init(void);
-
-/**
- * Get boot parameters
- */
-void *arch_get_boot_params(void);
-
-/**
- * Set boot parameters from architecture-specific entry code.
- */
-void arch_set_boot_params(void *params);
-
-/**
- * Get command line arguments
- */
-const char *arch_get_cmdline(void);
-
-/**
- * Get architecture name
- */
-const char *arch_get_name(void);
-
-/**
- * Get architecture bits (32/64)
- */
-uint32_t arch_get_bits(void);
-
-/**
- * Check if architecture supports specific feature
- */
-int arch_supports_feature(const char *feature);
-
-/**
- * Get architecture-specific compiler flags
- */
-const char *arch_get_cflags(void);
-
-/**
- * Get architecture-specific linker flags
- */
-const char *arch_get_ldflags(void);
+#include <ir0/arch_cpu.h>
