+++ includes/ir0/arch_io.h
@@ -1,0 +1,61 @@
+/* SPDX-License-Identifier: GPL-3.0-only */
+/**
+ * IR0 Kernel — Core system software
+ * Copyright (C) 2025  Iván Rodriguez
+ *
+ * This file is part of the IR0 Operating System.
+ * Distributed under the terms of the GNU General Public License v3.0.
+ * See the LICENSE file in the project root for full license information.
+ *
+ * File: arch_interface.h
+ * Description: IR0 kernel source/header file
+ */
+
+// arch/common/arch_interface.h
+#pragma once
+#include <stdint.h>
+
+
+void arch_enable_interrupts(void);
+void arch_disable_interrupts(void);
+
+
+uint8_t inb(uint16_t port);
+
+
+void outb(uint16_t port, uint8_t value);
+
+
+uintptr_t read_fault_address(void);
+
+
+const char *arch_get_name(void);
+
+
+void cpu_wait(void);
+void arch_cpu_idle(void);
+void arch_cpu_halt(void);
+void arch_system_halt(void) __attribute__((noreturn));
+void arch_system_reboot(void) __attribute__((noreturn));
+void arch_system_poweroff(void) __attribute__((noreturn));
+void arch_set_boot_params(void *params);
+
+
+
+#ifndef ARCH_X86_64
+#if defined(__x86_64__) || defined(__amd64__)
+#define ARCH_X86_64
+#elif defined(__i386__) || defined(__i486__) || defined(__i586__) || defined(__i686__)
+#define ARCH_X86_32
+#elif defined(__aarch64__)
+#ifndef ARCH_ARM64
+#define ARCH_ARM64
+#endif
+#elif defined(__arm__)
+#ifndef ARCH_ARM32
+#define ARCH_ARM32
+#endif
+#else
+#error "Arquitectura no soportada en arch_interface.h"
+#endif
+#endif /* ARCH_X86_64 */
